Calculate convective parameters

A core function for calculating convective parameters commonly used in the operational prediction of severe convective storms. Returns a vector of parameters.

sounding_compute(pressure, altitude, temp, dpt, wd, ws, accuracy = 2)

Arguments

pressure
pressure [hPa]

altitude
altitude [m] (can be above sea level or above ground level as function always consider first level as surface, i.e h = 0 m) altitude [metres]

temp
temperature [degree Celsius]

dpt
dew point temperature [degree Celsius]

wd
wind direction [azimuth in degrees]

ws
wind speed [knots]

accuracy
accuracy of computations where 3 = high (slow), 2 = medium (recommended), 1 = low (fast)

Value

Named vector of 100+ convective indices

MU_CAPE

MU_03km_CAPE

MU_HGL_CAPE

MU_CIN

MU_LCL_HGT

MU_LFC_HGT

MU_EL_HGT

MU_LI

MU_WMAX

MU_EL_TEMP

MU_LCL_TEMP

MU_LFC_TEMP

MU_MIXR

SB_CAPE

SB_03km_CAPE

SB_HGL_CAPE

SB_CIN

SB_LCL_HGT

SB_LFC_HGT

SB_EL_HGT

SB_LI

SB_WMAX

SB_EL_TEMP

SB_LCL_TEMP

SB_LFC_TEMP

SB_MIXR

ML_CAPE

ML_03km_CAPE

ML_HGL_CAPE

ML_CIN

ML_LCL_HGT

ML_LFC_HGT

ML_EL_HGT

ML_LI

ML_WMAX

ML_EL_TEMP

ML_LCL_TEMP

ML_LFC_TEMP

ML_MIXR

LR_01km

LR_03km

LR_24km

LR_36km

LR_500700hPa

LR_500800hPa

FRZG_HGT

FRZG_wetbulb_HGT

HGT_max_thetae_03km

HGT_min_thetae_04km

Delta_thetae

Delta_thetae_HGL

DCAPE

Cold_Pool_Strength

Wind_Index

PRCP_WATER

Moisture_Flux_02km

RH_02km

RH_25km

RH_HGL

BS_01km

BS_02km

BS_03km

BS_06km

BS_08km

BS_36km

BS_18km

BS_EFF_MU

BS_EFF_SB

BS_EFF_ML

BS_SFC_to_HGL

BS_MU_LFC_to_HGL

BS_SB_LFC_to_HGL

BS_ML_LFC_to_HGL

MW_01km

MW_02km

MW_06km

MW_13km

SRH_100m_RM

SRH_500m_RM

SRH_1km_RM

SRH_3km_RM

SRH_100m_LM

SRH_500m_LM

SRH_1km_LM

SRH_3km_LM

Bunkers_RM_A

Bunkers_RM_M

Bunkers_LM_A

Bunkers_LM_M

Bunkers_MW_A

Bunkers_MW_M

K_Index

Showalter_Index

TotalTotals_Index

SWEAT_Index

STP

STP_new

SCP

SCP_new

SHIP

DCP

MU_WMAXSHEAR

SB_WMAXSHEAR

ML_WMAXSHEAR

MU_EFF_WMAXSHEAR

SB_EFF_WMAXSHEAR

ML_EFF_WMAXSHEAR

Examples

pressure = c(1000, 855, 700, 500, 300, 100, 10)
altitude = c(0, 1500, 2500, 6000, 8500, 12000, 25000)
temp = c(25, 10, 0, -15, -30, -50, -92)
dpt = c(20, 5, -5, -30, -55, -80, -99)
wd = c(0, 90, 135, 180, 270, 350, 0)
ws = c(5, 10, 20, 30, 40, 5, 0)
accuracy = 2 
options(digits = 2) #change output formatting precision 
sounding_compute(pressure, altitude, temp, dpt, wd, ws, accuracy)

#>             MU_CAPE        MU_03km_CAPE         MU_HGL_CAPE              MU_CIN 
#>             2195.41              575.59             1568.01                0.00 
#>          MU_LCL_HGT          MU_LFC_HGT           MU_EL_HGT               MU_LI 
#>              730.00              730.00             8200.00               -9.63 
#>             MU_WMAX          MU_EL_TEMP         MU_LCL_TEMP         MU_LFC_TEMP 
#>               66.26              -28.20               17.70               17.70 
#>             MU_MIXR             SB_CAPE        SB_03km_CAPE         SB_HGL_CAPE 
#>               14.88             2195.41              575.59             1568.01 
#>              SB_CIN          SB_LCL_HGT          SB_LFC_HGT           SB_EL_HGT 
#>                0.00              730.00              730.00             8200.00 
#>               SB_LI             SB_WMAX          SB_EL_TEMP         SB_LCL_TEMP 
#>               -9.63               66.26              -28.20               17.70 
#>         SB_LFC_TEMP             SB_MIXR             ML_CAPE        ML_03km_CAPE 
#>               17.70               14.88             1559.41              416.88 
#>         ML_HGL_CAPE              ML_CIN          ML_LCL_HGT          ML_LFC_HGT 
#>             1187.96                0.00              975.00              975.00 
#>           ML_EL_HGT               ML_LI             ML_WMAX          ML_EL_TEMP 
#>             7800.00               -7.15               55.85              -25.80 
#>         ML_LCL_TEMP         ML_LFC_TEMP             ML_MIXR             LR_01km 
#>               15.25               15.25               13.02              -10.00 
#>             LR_03km             LR_24km             LR_36km        LR_500700hPa 
#>               -9.05               -5.77               -4.29               -4.29 
#>        LR_500800hPa            FRZG_HGT    FRZG_wetbulb_HGT HGT_max_thetae_03km 
#>               -6.67             2500.00             2300.00                0.00 
#> HGT_min_thetae_04km        Delta_thetae    Delta_thetae_HGL               DCAPE 
#>             3700.00               28.00               26.67              578.58 
#>  Cold_Pool_Strength          Wind_Index          PRCP_WATER  Moisture_Flux_02km 
#>              300.81               34.12               27.10               30.43 
#>             RH_02km             RH_25km              RH_HGL             BS_01km 
#>                0.72                0.56                0.46                3.83 
#>             BS_02km             BS_03km             BS_06km             BS_08km 
#>                8.78               12.66               18.01               17.41 
#>             BS_36km             BS_18km           BS_EFF_MU           BS_EFF_SB 
#>                9.37               20.28               14.14               14.14 
#>           BS_EFF_ML       BS_SFC_to_HGL    BS_MU_LFC_to_HGL    BS_SB_LFC_to_HGL 
#>               13.82               15.51               14.07               14.07 
#>    BS_ML_LFC_to_HGL             MW_01km             MW_02km             MW_06km 
#>               13.69                2.43                3.35                7.81 
#>             MW_13km         SRH_100m_RM         SRH_500m_RM          SRH_1km_RM 
#>                6.84                3.87               19.33               39.63 
#>          SRH_3km_RM         SRH_100m_LM         SRH_500m_LM          SRH_1km_LM 
#>              152.52                1.37                6.83               14.00 
#>          SRH_3km_LM        Bunkers_RM_A        Bunkers_RM_M        Bunkers_LM_A 
#>              -13.13              209.40                7.79              122.06 
#>        Bunkers_LM_M        Bunkers_MW_A        Bunkers_MW_M             K_Index 
#>               13.18              151.95                7.81               24.35 
#>     Showalter_Index   TotalTotals_Index         SWEAT_Index                 STP 
#>                3.90               44.35              106.42                0.35 
#>             STP_new                 SCP             SCP_new                SHIP 
#>                0.19                6.03                4.73                0.61 
#>                 DCP        MU_WMAXSHEAR        SB_WMAXSHEAR        ML_WMAXSHEAR 
#>                1.08             1193.11             1193.11             1005.54 
#>    MU_EFF_WMAXSHEAR    SB_EFF_WMAXSHEAR    ML_EFF_WMAXSHEAR 
#>              936.94              936.94              771.71