R call to C++ function for calculating thermo- and kinematic indices derived from atmospheric profiling.

More details in the sounding_compute() function

sounding_default(
  pressure,
  altitude,
  temp,
  dpt,
  wd,
  ws,
  export_profile,
  accuracy,
  interpolate_step,
  meanlayer_bottom_top,
  storm_motion
)

Arguments

pressure

pressure [hPa]

altitude

altitude [meters]

temp

temperature [degree Celsius]

dpt

dew point temperature [degree Celsius]

wd

wind direction [azimuth in degrees]

ws

wind speed [knots]

export_profile

possibility to export interpolated profile on the levels defined in accuracy configuration

accuracy

accuracy of computations where 3 = high (slow), 2 = medium (recommended), 1 = low (fast)

interpolate_step

interpolation step to be used for vertical interpolation. Valid only if `accuracy` is set to 3 (default is 5 m)

meanlayer_bottom_top

(optional) vector of length 2 for bottom and top heights used for computing parcel starting parameters; default: 0, 500

storm_motion

(optional) for moving storms only - one can define vector of length two with wind speed (m/s) and wind directions (degrees) that will be used to compute adjusted SRH parameters

Value

MU_CAPE

MU_CAPE_M10

MU_CAPE_M10_PT

MU_02km_CAPE

MU_03km_CAPE

MU_HGL_CAPE

MU_CIN

MU_LCL_HGT

MU_LFC_HGT

MU_EL_HGT

MU_LI

MU_LI_M10

MU_WMAX

MU_EL_TEMP

MU_LCL_TEMP

MU_LFC_TEMP

MU_MIXR

MU_CAPE_500

MU_CAPE_500_M10

MU_CAPE_500_M10_PT

MU_CIN_500

MU_LI_500

MU_LI_500_M10

SB_CAPE

SB_CAPE_M10

SB_CAPE_M10_PT

SB_02km_CAPE

SB_03km_CAPE

SB_HGL_CAPE

SB_CIN

SB_LCL_HGT

SB_LFC_HGT

SB_EL_HGT

SB_LI

SB_LI_M10

SB_WMAX

SB_EL_TEMP

SB_LCL_TEMP

SB_LFC_TEMP

SB_MIXR

ML_CAPE

ML_CAPE_M10

ML_CAPE_M10_PT

ML_02km_CAPE

ML_03km_CAPE

ML_HGL_CAPE

ML_CIN

ML_LCL_HGT

ML_LFC_HGT

ML_EL_HGT

ML_LI

ML_LI_M10

ML_WMAX

ML_EL_TEMP

ML_LCL_TEMP

ML_LFC_TEMP

ML_MIXR

LR_0500m

LR_01km

LR_02km

LR_03km

LR_04km

LR_06km

LR_16km

LR_26km

LR_24km

LR_36km

LR_26km_MAX

LR_500700hPa

LR_500800hPa

LR_600800hPa

FRZG_HGT

FRZG_wetbulb_HGT

HGT_max_thetae_03km

HGT_min_thetae_04km

Delta_thetae

Delta_thetae_min04km

Thetae_01km

Thetae_02km

DCAPE

Cold_Pool_Strength

Wind_Index

PRCP_WATER

Moisture_Flux_02km

RH_01km

RH_02km

RH_14km

RH_25km

RH_36km

RH_HGL

BS_0500m

BS_01km

BS_02km

BS_03km

BS_06km

BS_08km

BS_36km

BS_26km

BS_16km

BS_18km

BS_EFF_MU

BS_EFF_SB

BS_EFF_ML

BS_SFC_to_M10

BS_1km_to_M10

BS_2km_to_M10

BS_MU_LFC_to_M10

BS_SB_LFC_to_M10

BS_ML_LFC_to_M10

BS_MW02_to_SM

BS_MW02_to_RM

BS_MW02_to_LM

BS_HGL_to_SM

BS_HGL_to_RM

BS_HGL_to_LM

MW_0500m

MW_01km

MW_02km

MW_03km

MW_06km

MW_13km

SRH_100m_RM

SRH_250m_RM

SRH_500m_RM

SRH_1km_RM

SRH_3km_RM

SRH_36km_RM

SRH_100m_LM

SRH_250m_LM

SRH_500m_LM

SRH_1km_LM

SRH_3km_LM

SRH_36km_LM

SV_500m_RM

SV_01km_RM

SV_03km_RM

SV_500m_LM

SV_01km_LM

SV_03km_LM

MW_SR_500m_RM

MW_SR_01km_RM

MW_SR_03km_RM

MW_SR_500m_LM

MW_SR_01km_LM

MW_SR_03km_LM

MW_SR_VM_500m_RM

MW_SR_VM_01km_RM

MW_SR_VM_03km_RM

MW_SR_VM_500m_LM

MW_SR_VM_01km_LM

MW_SR_VM_03km_LM

SV_FRA_500m_RM

SV_FRA_01km_RM

SV_FRA_03km_RM

SV_FRA_500m_LM

SV_FRA_01km_LM

SV_FRA_03km_LM

Bunkers_RM_A

Bunkers_RM_M

Bunkers_LM_A

Bunkers_LM_M

Bunkers_MW_A

Bunkers_MW_M

Corfidi_downwind_A

Corfidi_downwind_M

Corfidi_upwind_A

Corfidi_upwind_M

K_Index

Showalter_Index

TotalTotals_Index

SWEAT_Index

STP_fix

STP_new

STP_fix_LM

STP_new_LM

SCP_fix

SCP_new

SCP_fix_LM

SCP_new_LM

SHIP

HSI

DCP

MU_WMAXSHEAR

SB_WMAXSHEAR

ML_WMAXSHEAR

MU_EFF_WMAXSHEAR

SB_EFF_WMAXSHEAR

ML_EFF_WMAXSHEAR

EHI_500m

EHI_01km

EHI_03km

EHI_500m_LM

EHI_01km_LM

EHI_03km_LM

SHERBS3

SHERBE

SHERBS3_v2

SHERBE_v2

DEI

DEI_eff

TIP

Examples

pressure = c(1000, 855, 700, 500, 300, 100, 10) 
altitude = c(0, 1500, 2500, 6000, 8500, 12000, 25000)
temp = c(25, 10, 0, -15, -30, -50, -92)
dpt = c(20, 5, -5, -30, -55, -80, -99)
wd = c(0, 90, 135, 180, 270, 350, 0)
ws = c(5, 10, 20, 30, 40, 5, 0)
sounding_default(pressure, altitude, temp, dpt, wd, ws,
                 accuracy = 3,
                 export_profile = 0,
                 interpolate_step = 5,
                 storm_motion = c(999, 999),
                 meanlayer_bottom_top = c(0, 500))
#>   [1]  2.284081e+03  1.004494e+03  3.221361e+02  2.512952e+02  5.795603e+02
#>   [6]  1.621657e+03  0.000000e+00  7.300000e+02  7.300000e+02  8.245000e+03
#>  [11] -1.011193e+01 -1.085707e+01  6.758817e+01 -2.847000e+01  1.770000e+01
#>  [16]  1.770000e+01  1.487592e+01  1.065991e+03  3.836773e+02  1.428513e+02
#>  [21]  0.000000e+00 -4.953444e+00 -6.162641e+00  2.284081e+03  1.004494e+03
#>  [26]  3.221361e+02  2.512952e+02  5.795603e+02  1.621657e+03  0.000000e+00
#>  [31]  7.300000e+02  7.300000e+02  8.245000e+03 -1.011193e+01 -1.085707e+01
#>  [36]  6.758817e+01 -2.847000e+01  1.770000e+01  1.770000e+01  1.487592e+01
#>  [41]  1.642718e+03  6.673561e+02  2.309205e+02  1.651221e+02  4.223024e+02
#>  [46]  1.245385e+03  0.000000e+00  9.600000e+02  9.600000e+02  7.820000e+03
#>  [51] -7.558895e+00 -8.534296e+00  5.731872e+01 -2.592000e+01  1.540000e+01
#>  [56]  1.540000e+01  1.304875e+01 -1.000000e+01 -1.000000e+01 -1.000000e+01
#>  [61] -9.047619e+00 -7.857143e+00 -6.666667e+00 -6.000000e+00 -5.000000e+00
#>  [66] -5.724973e+00 -4.285714e+00 -5.714286e+00 -4.285714e+00 -5.174910e+00
#>  [71] -5.824635e+00  2.500000e+03  2.265000e+03  0.000000e+00  3.745000e+03
#>  [76]  2.806979e+01  2.883458e+01  3.305323e+02  3.236191e+02  6.168139e+02
#>  [81]  1.270725e+01  3.382169e+01  2.710398e+01  3.042548e+01  7.291035e-01
#>  [86]  7.197069e-01  6.452043e-01  5.550179e-01  4.436036e-01  4.117969e-01
#>  [91]  1.917220e+00  3.834439e+00  8.782108e+00  1.265597e+01  1.800554e+01
#>  [96]  1.740768e+01  9.369290e+00  1.333042e+01  1.664783e+01  2.027907e+01
#> [101]  1.417295e+01  1.417295e+01  1.383392e+01  1.548068e+01  1.361265e+01
#> [106]  9.838748e+00  1.403852e+01  1.403852e+01  1.367164e+01  7.304028e+00
#> [111]  1.410166e+03  1.410166e+03  4.893412e+00  1.419401e+03  1.419401e+03
#> [116]  2.308638e+00  2.425112e+00  3.347647e+00  4.800332e+00  7.810701e+00
#> [121]  6.838887e+00 -5.386997e+02 -1.308271e+03 -2.590889e+03 -5.156126e+03
#> [126] -1.716614e+04 -1.775201e+04 -5.386997e+02 -1.308271e+03 -2.590889e+03
#> [131] -5.156126e+03 -1.716614e+04 -1.775201e+04 -3.673263e-03 -3.655864e-03
#> [136] -4.056463e-03 -3.673263e-03 -3.655864e-03 -4.056463e-03  1.410678e+03
#> [141]  1.410374e+03  1.411080e+03  1.410678e+03  1.410374e+03  1.411080e+03
#> [146]  1.410675e+03  1.410372e+03  1.411089e+03  1.410675e+03  1.410372e+03
#> [151]  1.411089e+03 -9.484812e-01 -9.486851e-01 -8.162005e-01 -9.484812e-01
#> [156] -9.486851e-01 -8.162005e-01  4.500000e+01  1.412799e+03  4.500000e+01
#> [161]  1.412799e+03  1.519494e+02  7.810701e+00  2.186955e+02  1.469824e+01
#> [166]  2.313283e+02  9.179441e+00  2.435484e+01  3.841794e+00  4.435484e+01
#> [171]  1.064168e+02 -4.712253e+01 -2.616825e+01 -4.712253e+01 -2.616825e+01
#> [176] -7.059763e+02 -5.557047e+02 -7.059763e+02 -5.557047e+02  6.325821e-01
#> [181]  1.707597e+00  1.193655e+00  1.216962e+03  1.216962e+03  1.032055e+03
#> [186]  9.579236e+02  9.579236e+02  7.929429e+02 -3.698625e+01 -7.360630e+01
#> [191] -2.450552e+02 -3.698625e+01 -7.360630e+01 -2.450552e+02  6.481689e-01
#> [196]  6.989765e-01  8.642252e-01  9.319686e-01  1.536380e+00  1.199631e+00
#> [201] -9.105982e+01