Save `sounding_layout` to a graphical file

Auxiliary function to `sounding_plot` that plots a composite \ of Skew-T, hodograph and selected convective parameters \ on a single layout and saves as graphical file.

sounding_save(
  pressure,
  altitude,
  temp,
  dpt,
  wd,
  ws,
  title = "",
  parcel = "MU",
  max_speed = 25,
  buoyancy_polygon = TRUE,
  SRH_polygon = "03km",
  DCAPE = FALSE,
  filename,
  ...
)

Arguments

pressure

pressure [hPa]

altitude

altitude [m] (can be above sea level or above ground level as function always consider first level as a surface, i.e h = 0 m)

temp

temperature [degree Celsius]

dpt

dew point temperature [degree Celsius]

wd

wind direction in degrees [azimuth in degrees]

ws

wind speed [knots]

title

title to be added in the layout's header

parcel

parcel tracing on Skew-T for "MU", "ML" or "SB" parcel, "none" for no parcel line.

max_speed

range of the hodograph to be drawn, 25 m/s used as default

buoyancy_polygon

logical, plotting area of parcel's positive (yellow) or negative (red) buoyancy (default = TRUE)

SRH_polygon

draws polygon for storm-relative helicity, available options are "0500m", "01km", "03km", "36km", "none", "03km" used as default

DCAPE

draws downdraft parcel and polygon of downdraft's negative buoyancy (default = FALSE)

filename

output file name with extension indicating file format (e.g. "my_plot.png" or "my_plot.svg")

...

other arguments that can be used with `sounding_plot` or other graphic arguments

Value

graphical file with Skew-T and hodograph on a single layout

Examples

# \donttest{
data("sounding_vienna")
attach(sounding_vienna)
#> The following objects are masked from sounding_vienna (pos = 3):
#> 
#>     altitude, dpt, pressure, temp, wd, ws
#> The following objects are masked from sounding_vienna (pos = 4):
#> 
#>     altitude, dpt, pressure, temp, wd, ws
#> The following objects are masked from sounding_vienna (pos = 5):
#> 
#>     altitude, dpt, pressure, temp, wd, ws
#> The following objects are masked from sounding_vienna (pos = 6):
#> 
#>     altitude, dpt, pressure, temp, wd, ws
#> The following object is masked from package:datasets:
#> 
#>     pressure
sounding_save(filename = tempfile(), 
             pressure, altitude, temp, dpt, wd, ws, parcel = "MU", 
             title = "Vienna - 23 August 2011, 12:00 UTC")
# }