Calculate convective parameters

A core function for calculating convective parameters commonly used in the operational prediction of severe convective storms. Returns a vector of parameters.

sounding_compute(
  pressure,
  altitude,
  temp,
  dpt,
  wd,
  ws,
  accuracy = 2,
  interpolate_step = 5,
  meanlayer_bottom_top = c(0, 500),
  storm_motion = c(999, 999)
)

Arguments

pressure

pressure [hPa]

altitude

altitude [m] (can be above sea level or above ground level as function always consider first level as surface, i.e h = 0 m) altitude [metres]

temp

temperature [degree Celsius]

dpt

dew point temperature [degree Celsius]

wd

wind direction [azimuth in degrees]

ws

wind speed [knots]

accuracy

accuracy of computations where 3 = high (slow), 2 = medium (recommended), 1 = low (fast)

interpolate_step

interpolation step to be used for vertical interpolation. Valid only if `accuracy` is set to 3 (default is 5 m)

meanlayer_bottom_top

(optional) vector of length 2 for bottom and top heights used for computing parcel starting parameters; default: 0, 500

storm_motion

(optional) for moving storms only - one can define vector of length two with wind speed (m/s) and wind directions (degrees) that will be used to compute adjusted SRH parameters

Value

Named vector of 200+ convective indices

Details

MU_CAPE

MU_CAPE_M10

MU_CAPE_M10_PT

MU_02km_CAPE

MU_03km_CAPE

MU_HGL_CAPE

MU_CIN

MU_LCL_HGT

MU_LFC_HGT

MU_EL_HGT

MU_LI

MU_LI_M10

MU_WMAX

MU_EL_TEMP

MU_LCL_TEMP

MU_LFC_TEMP

MU_MIXR

MU_CAPE_500

MU_CAPE_500_M10

MU_CAPE_500_M10_PT

MU_CIN_500

MU_LI_500

MU_LI_500_M10

SB_CAPE

SB_CAPE_M10

SB_CAPE_M10_PT

SB_02km_CAPE

SB_03km_CAPE

SB_HGL_CAPE

SB_CIN

SB_LCL_HGT

SB_LFC_HGT

SB_EL_HGT

SB_LI

SB_LI_M10

SB_WMAX

SB_EL_TEMP

SB_LCL_TEMP

SB_LFC_TEMP

SB_MIXR

ML_CAPE

ML_CAPE_M10

ML_CAPE_M10_PT

ML_02km_CAPE

ML_03km_CAPE

ML_HGL_CAPE

ML_CIN

ML_LCL_HGT

ML_LFC_HGT

ML_EL_HGT

ML_LI

ML_LI_M10

ML_WMAX

ML_EL_TEMP

ML_LCL_TEMP

ML_LFC_TEMP

ML_MIXR

LR_0500m

LR_01km

LR_02km

LR_03km

LR_04km

LR_06km

LR_16km

LR_26km

LR_24km

LR_36km

LR_26km_MAX

LR_500700hPa

LR_500800hPa

LR_600800hPa

FRZG_HGT

FRZG_wetbulb_HGT

HGT_max_thetae_03km

HGT_min_thetae_04km

Delta_thetae

Delta_thetae_min04km

Thetae_01km

Thetae_02km

DCAPE

Cold_Pool_Strength

Wind_Index

PRCP_WATER

Moisture_Flux_02km

RH_01km

RH_02km

RH_14km

RH_25km

RH_36km

RH_HGL

BS_0500m

BS_01km

BS_02km

BS_03km

BS_06km

BS_08km

BS_36km

BS_26km

BS_16km

BS_18km

BS_EFF_MU

BS_EFF_SB

BS_EFF_ML

BS_SFC_to_M10

BS_1km_to_M10

BS_2km_to_M10

BS_MU_LFC_to_M10

BS_SB_LFC_to_M10

BS_ML_LFC_to_M10

BS_MW02_to_SM

BS_MW02_to_RM

BS_MW02_to_LM

BS_HGL_to_SM

BS_HGL_to_RM

BS_HGL_to_LM

MW_0500m

MW_01km

MW_02km

MW_03km

MW_06km

MW_13km

SRH_100m_RM

SRH_250m_RM

SRH_500m_RM

SRH_1km_RM

SRH_3km_RM

SRH_36km_RM

SRH_100m_LM

SRH_250m_LM

SRH_500m_LM

SRH_1km_LM

SRH_3km_LM

SRH_36km_LM

SV_500m_RM

SV_01km_RM

SV_03km_RM

SV_500m_LM

SV_01km_LM

SV_03km_LM

MW_SR_500m_RM

MW_SR_01km_RM

MW_SR_03km_RM

MW_SR_500m_LM

MW_SR_01km_LM

MW_SR_03km_LM

MW_SR_VM_500m_RM

MW_SR_VM_01km_RM

MW_SR_VM_03km_RM

MW_SR_VM_500m_LM

MW_SR_VM_01km_LM

MW_SR_VM_03km_LM

SV_FRA_500m_RM

SV_FRA_01km_RM

SV_FRA_03km_RM

SV_FRA_500m_LM

SV_FRA_01km_LM

SV_FRA_03km_LM

Bunkers_RM_A

Bunkers_RM_M

Bunkers_LM_A

Bunkers_LM_M

Bunkers_MW_A

Bunkers_MW_M

Corfidi_downwind_A

Corfidi_downwind_M

Corfidi_upwind_A

Corfidi_upwind_M

K_Index

Showalter_Index

TotalTotals_Index

SWEAT_Index

STP_fix

STP_new

STP_fix_LM

STP_new_LM

SCP_fix

SCP_new

SCP_fix_LM

SCP_new_LM

SHIP

HSI

DCP

MU_WMAXSHEAR

SB_WMAXSHEAR

ML_WMAXSHEAR

MU_EFF_WMAXSHEAR

SB_EFF_WMAXSHEAR

ML_EFF_WMAXSHEAR

EHI_500m

EHI_01km

EHI_03km

EHI_500m_LM

EHI_01km_LM

EHI_03km_LM

SHERBS3

SHERBE

SHERBS3_v2

SHERBE_v2

DEI

DEI_eff

TIP

Examples

old_options = options(scipen = 99) 
pressure = c(1000, 855, 700, 500, 300, 100, 10)
altitude = c(0, 1500, 2500, 6000, 8500, 12000, 25000)
temp = c(25, 10, 0, -15, -30, -50, -92)
dpt = c(20, 5, -5, -30, -55, -80, -99)
wd = c(0, 90, 135, 180, 270, 350, 0)
ws = c(5, 10, 20, 30, 40, 5, 0)
accuracy = 2
sounding_compute(pressure, altitude, temp, dpt, wd, ws, accuracy)
#>              MU_CAPE          MU_CAPE_M10       MU_CAPE_M10_PT 
#>       2269.925733301        998.144308627        313.074714966 
#>         MU_02km_CAPE         MU_03km_CAPE          MU_HGL_CAPE 
#>        247.979380820        575.629305978       1616.538411895 
#>               MU_CIN           MU_LCL_HGT           MU_LFC_HGT 
#>          0.000000000        730.000000000        730.000000000 
#>            MU_EL_HGT                MU_LI            MU_LI_M10 
#>       8300.000000000        -10.111930240        -10.853947310 
#>              MU_WMAX           MU_EL_TEMP          MU_LCL_TEMP 
#>         67.378419888        -28.800000000         17.700000000 
#>          MU_LFC_TEMP              MU_MIXR          MU_CAPE_500 
#>         17.700000000         14.875917910       1076.032178138 
#>      MU_CAPE_500_M10   MU_CAPE_500_M10_PT           MU_CIN_500 
#>        389.365140754        137.081449189          0.000000000 
#>            MU_LI_500        MU_LI_500_M10              SB_CAPE 
#>         -5.041702392         -6.234631722       2269.925733301 
#>          SB_CAPE_M10       SB_CAPE_M10_PT         SB_02km_CAPE 
#>        998.144308627        313.074714966        247.979380820 
#>         SB_03km_CAPE          SB_HGL_CAPE               SB_CIN 
#>        575.629305978       1616.538411895          0.000000000 
#>           SB_LCL_HGT           SB_LFC_HGT            SB_EL_HGT 
#>        730.000000000        730.000000000       8300.000000000 
#>                SB_LI            SB_LI_M10              SB_WMAX 
#>        -10.111930240        -10.853947310         67.378419888 
#>           SB_EL_TEMP          SB_LCL_TEMP          SB_LFC_TEMP 
#>        -28.800000000         17.700000000         17.700000000 
#>              SB_MIXR              ML_CAPE          ML_CAPE_M10 
#>         14.875917910       1646.063927273        670.100095437 
#>       ML_CAPE_M10_PT         ML_02km_CAPE         ML_03km_CAPE 
#>        225.281641180        164.079828657        422.429000173 
#>          ML_HGL_CAPE               ML_CIN           ML_LCL_HGT 
#>       1250.022127396          0.000000000        975.000000000 
#>           ML_LFC_HGT            ML_EL_HGT                ML_LI 
#>        975.000000000       7900.000000000         -7.620329500 
#>            ML_LI_M10              ML_WMAX           ML_EL_TEMP 
#>         -8.584526572         57.377067323        -26.400000000 
#>          ML_LCL_TEMP          ML_LFC_TEMP              ML_MIXR 
#>         15.250000000         15.250000000         13.048747763 
#>             LR_0500m              LR_01km              LR_02km 
#>        -10.000000000        -10.000000000        -10.000000000 
#>              LR_03km              LR_04km              LR_06km 
#>         -9.047619048         -7.857142857         -6.666666667 
#>              LR_16km              LR_26km              LR_24km 
#>         -6.000000000         -5.000000000         -5.767195767 
#>              LR_36km          LR_26km_MAX         LR_500700hPa 
#>         -4.285714286         -5.714285714         -4.285714286 
#>         LR_500800hPa         LR_600800hPa             FRZG_HGT 
#>         -5.180722892         -5.833333333       2500.000000000 
#>     FRZG_wetbulb_HGT  HGT_max_thetae_03km  HGT_min_thetae_04km 
#>       2275.000000000          0.000000000       3750.000000000 
#>         Delta_thetae Delta_thetae_min04km          Thetae_01km 
#>         28.069785504         28.834554624        330.532272184 
#>          Thetae_02km                DCAPE   Cold_Pool_Strength 
#>        323.619051941        598.310049405         12.632184321 
#>           Wind_Index           PRCP_WATER   Moisture_Flux_02km 
#>         33.906407176         27.104631918         30.425480722 
#>              RH_01km              RH_02km              RH_14km 
#>          0.729103473          0.719706866          0.645204350 
#>              RH_25km              RH_36km               RH_HGL 
#>          0.555017886          0.443603638          0.460317694 
#>             BS_0500m              BS_01km              BS_02km 
#>          1.917219591          3.834439182          8.782108409 
#>              BS_03km              BS_06km              BS_08km 
#>         12.655968060         18.005540000         17.407678201 
#>              BS_36km              BS_26km              BS_16km 
#>          9.369289559         13.330420492         16.647826116 
#>              BS_18km            BS_EFF_MU            BS_EFF_SB 
#>         20.279072460         14.223187917         14.223187917 
#>            BS_EFF_ML        BS_SFC_to_M10        BS_1km_to_M10 
#>         13.896750821         15.510376997         13.649859352 
#>        BS_2km_to_M10     BS_MU_LFC_to_M10     BS_SB_LFC_to_M10 
#>          9.883032095         14.073701016         14.073701016 
#>     BS_ML_LFC_to_M10        BS_MW02_to_SM        BS_MW02_to_RM 
#>         13.686365658          7.304027597         10.140964728 
#>        BS_MW02_to_LM         BS_HGL_to_SM         BS_HGL_to_RM 
#>         10.786958453          4.893411960          7.785950290 
#>         BS_HGL_to_LM             MW_0500m              MW_01km 
#>          9.988490342          2.308638103          2.425112273 
#>              MW_02km              MW_03km              MW_06km 
#>          3.347646594          4.800331697          7.810700539 
#>              MW_13km          SRH_100m_RM          SRH_250m_RM 
#>          6.838886565          4.253473515         10.053664673 
#>          SRH_500m_RM           SRH_1km_RM           SRH_3km_RM 
#>         19.720649935         39.634639574        152.521946573 
#>          SRH_36km_RM          SRH_100m_LM          SRH_250m_LM 
#>        236.590130791          1.502688386          3.551808913 
#>          SRH_500m_LM           SRH_1km_LM           SRH_3km_LM 
#>          6.967009791         14.002323601        -13.130823085 
#>          SRH_36km_LM           SV_500m_RM           SV_01km_RM 
#>        -24.378976558          0.003904164          0.003879744 
#>           SV_03km_RM           SV_500m_LM           SV_01km_LM 
#>          0.004796086          0.001013916          0.001100521 
#>           SV_03km_LM        MW_SR_500m_RM        MW_SR_01km_RM 
#>         -0.001435868         10.086252358         10.150075094 
#>        MW_SR_03km_RM        MW_SR_500m_LM        MW_SR_01km_LM 
#>          9.535939066         13.782126361         12.858486067 
#>        MW_SR_03km_LM     MW_SR_VM_500m_RM     MW_SR_VM_01km_RM 
#>          8.457941382         10.107849533         10.225339777 
#>     MW_SR_VM_03km_RM     MW_SR_VM_500m_LM     MW_SR_VM_01km_LM 
#>         10.535820052         13.764654240         12.837090856 
#>     MW_SR_VM_03km_LM       SV_FRA_500m_RM       SV_FRA_01km_RM 
#>          8.834244774          0.998219354          0.987136746 
#>       SV_FRA_03km_RM       SV_FRA_500m_LM       SV_FRA_01km_LM 
#>          0.956008752          0.259238846          0.280009490 
#>       SV_FRA_03km_LM         Bunkers_RM_A         Bunkers_RM_M 
#>         -0.286213018        209.404566875          7.793284819 
#>         Bunkers_LM_A         Bunkers_LM_M         Bunkers_MW_A 
#>        122.058537262         13.182518634        151.949436108 
#>         Bunkers_MW_M   Corfidi_downwind_A   Corfidi_downwind_M 
#>          7.810700539        218.695496108         14.698235841 
#>     Corfidi_upwind_A     Corfidi_upwind_M              K_Index 
#>        231.328321825          9.179441145         24.354838710 
#>      Showalter_Index    TotalTotals_Index          SWEAT_Index 
#>          3.750147453         44.354838710        106.416793661 
#>              STP_fix              STP_new           STP_fix_LM 
#>          0.359981513          0.200492750          0.127176069 
#>           STP_new_LM              SCP_fix              SCP_new 
#>          0.070831081          6.233760868          4.924259548 
#>           SCP_fix_LM           SCP_new_LM                 SHIP 
#>         -0.536673003         -0.423936243          0.628661858 
#>                  HSI                  DCP         MU_WMAXSHEAR 
#>          1.715903102          1.150670745       1213.184834430 
#>         SB_WMAXSHEAR         ML_WMAXSHEAR     MU_EFF_WMAXSHEAR 
#>       1213.184834430       1033.105080761        958.335927637 
#>     SB_EFF_WMAXSHEAR     ML_EFF_WMAXSHEAR             EHI_500m 
#>        958.335927637        797.354807442          0.279777567 
#>             EHI_01km             EHI_03km          EHI_500m_LM 
#>          0.562298052          2.163834321          0.098841218 
#>          EHI_01km_LM          EHI_03km_LM              SHERBS3 
#>          0.198651467         -0.186287458          0.648168875 
#>               SHERBE           SHERBS3_v2            SHERBE_v2 
#>          0.701454223          0.864225167          0.935272297 
#>                  DEI              DEI_eff                  TIP 
#>          1.519761039          1.188457460          2.435630415 
options(old_options)