Calculate convective parameters

A core function for calculating convective parameters commonly used in the operational prediction of severe convective storms. Returns a vector of parameters.

sounding_compute(
  pressure,
  altitude,
  temp,
  dpt,
  wd,
  ws,
  accuracy = 2,
  interpolate_step = 5,
  meanlayer_bottom_top = c(0, 500),
  storm_motion = c(999, 999)
)

Arguments

pressure

pressure [hPa]

altitude

altitude [m] (can be above sea level or above ground level as function always consider first level as surface, i.e h = 0 m) altitude [metres]

temp

temperature [degree Celsius]

dpt

dew point temperature [degree Celsius]

wd

wind direction [azimuth in degrees]

ws

wind speed [knots]

accuracy

accuracy of computations where 3 = high (slow), 2 = medium (recommended), 1 = low (fast)

interpolate_step

interpolation step to be used for vertical interpolation. Valid only if `accuracy` is set to 3 (default is 5 m)

meanlayer_bottom_top

(optional) vector of length 2 for bottom and top heights used for computing parcel starting parameters; default: 0, 500

storm_motion

(optional) for moving storms only - one can define wind speed and wind directions (TODO: units!!!) that will be used to compute adjusted SRH parameters

Value

Named vector of 200+ convective indices

Details

MU_CAPE

MU_CAPE_M10

MU_CAPE_M10_PT

MU_02km_CAPE

MU_03km_CAPE

MU_HGL_CAPE

MU_CIN

MU_LCL_HGT

MU_LFC_HGT

MU_EL_HGT

MU_LI

MU_LI_M10

MU_WMAX

MU_EL_TEMP

MU_LCL_TEMP

MU_LFC_TEMP

MU_MIXR

MU_CAPE_500

MU_CAPE_500_M10

MU_CAPE_500_M10_PT

MU_CIN_500

MU_LI_500

MU_LI_500_M10

SB_CAPE

SB_CAPE_M10

SB_CAPE_M10_PT

SB_02km_CAPE

SB_03km_CAPE

SB_HGL_CAPE

SB_CIN

SB_LCL_HGT

SB_LFC_HGT

SB_EL_HGT

SB_LI

SB_LI_M10

SB_WMAX

SB_EL_TEMP

SB_LCL_TEMP

SB_LFC_TEMP

SB_MIXR

ML_CAPE

ML_CAPE_M10

ML_CAPE_M10_PT

ML_02km_CAPE

ML_03km_CAPE

ML_HGL_CAPE

ML_CIN

ML_LCL_HGT

ML_LFC_HGT

ML_EL_HGT

ML_LI

ML_LI_M10

ML_WMAX

ML_EL_TEMP

ML_LCL_TEMP

ML_LFC_TEMP

ML_MIXR

LR_0500m

LR_01km

LR_02km

LR_03km

LR_04km

LR_06km

LR_16km

LR_26km

LR_24km

LR_36km

LR_26km_MAX

LR_500700hPa

LR_500800hPa

LR_600800hPa

FRZG_HGT

FRZG_wetbulb_HGT

HGT_max_thetae_03km

HGT_min_thetae_04km

Delta_thetae

Delta_thetae_min04km

Thetae_01km

Thetae_02km

DCAPE

Cold_Pool_Strength

Wind_Index

PRCP_WATER

Moisture_Flux_02km

RH_01km

RH_02km

RH_14km

RH_25km

RH_36km

RH_HGL

BS_0500m

BS_01km

BS_02km

BS_03km

BS_06km

BS_08km

BS_36km

BS_26km

BS_16km

BS_18km

BS_EFF_MU

BS_EFF_SB

BS_EFF_ML

BS_SFC_to_M10

BS_1km_to_M10

BS_2km_to_M10

BS_MU_LFC_to_M10

BS_SB_LFC_to_M10

BS_ML_LFC_to_M10

BS_MW02_to_SM

BS_MW02_to_RM

BS_MW02_to_LM

BS_HGL_to_SM

BS_HGL_to_RM

BS_HGL_to_LM

MW_0500m

MW_01km

MW_02km

MW_03km

MW_06km

MW_13km

SRH_100m_RM

SRH_250m_RM

SRH_500m_RM

SRH_1km_RM

SRH_3km_RM

SRH_36km_RM

SRH_100m_LM

SRH_250m_LM

SRH_500m_LM

SRH_1km_LM

SRH_3km_LM

SRH_36km_LM

SV_500m_RM

SV_01km_RM

SV_03km_RM

SV_500m_LM

SV_01km_LM

SV_03km_LM

MW_SR_500m_RM

MW_SR_01km_RM

MW_SR_03km_RM

MW_SR_500m_LM

MW_SR_01km_LM

MW_SR_03km_LM

MW_SR_VM_500m_RM

MW_SR_VM_01km_RM

MW_SR_VM_03km_RM

MW_SR_VM_500m_LM

MW_SR_VM_01km_LM

MW_SR_VM_03km_LM

SV_FRA_500m_RM

SV_FRA_01km_RM

SV_FRA_03km_RM

SV_FRA_500m_LM

SV_FRA_01km_LM

SV_FRA_03km_LM

Bunkers_RM_A

Bunkers_RM_M

Bunkers_LM_A

Bunkers_LM_M

Bunkers_MW_A

Bunkers_MW_M

Corfidi_downwind_A

Corfidi_downwind_M

Corfidi_upwind_A

Corfidi_upwind_M

K_Index

Showalter_Index

TotalTotals_Index

SWEAT_Index

STP_fix

STP_new

STP_fix_LM

STP_new_LM

SCP_fix

SCP_new

SCP_fix_LM

SCP_new_LM

SHIP

HSI

DCP

MU_WMAXSHEAR

SB_WMAXSHEAR

ML_WMAXSHEAR

MU_EFF_WMAXSHEAR

SB_EFF_WMAXSHEAR

ML_EFF_WMAXSHEAR

EHI_500m

EHI_01km

EHI_03km

EHI_500m_LM

EHI_01km_LM

EHI_03km_LM

SHERBS3

SHERBE

SHERBS3_v2

SHERBE_v2

DEI

DEI_eff

TIP

Examples

pressure = c(1000, 855, 700, 500, 300, 100, 10)
altitude = c(0, 1500, 2500, 6000, 8500, 12000, 25000)
temp = c(25, 10, 0, -15, -30, -50, -92)
dpt = c(20, 5, -5, -30, -55, -80, -99)
wd = c(0, 90, 135, 180, 270, 350, 0)
ws = c(5, 10, 20, 30, 40, 5, 0)
accuracy = 2 
options(digits = 2) #change output formatting precision 
sounding_compute(pressure, altitude, temp, dpt, wd, ws, accuracy)
#>              MU_CAPE          MU_CAPE_M10       MU_CAPE_M10_PT 
#>              2.3e+03              1.0e+03              3.1e+02 
#>         MU_02km_CAPE         MU_03km_CAPE          MU_HGL_CAPE 
#>              2.5e+02              5.8e+02              1.6e+03 
#>               MU_CIN           MU_LCL_HGT           MU_LFC_HGT 
#>              0.0e+00              7.3e+02              7.3e+02 
#>            MU_EL_HGT                MU_LI            MU_LI_M10 
#>              8.3e+03             -1.0e+01             -1.1e+01 
#>              MU_WMAX           MU_EL_TEMP          MU_LCL_TEMP 
#>              6.7e+01             -2.9e+01              1.8e+01 
#>          MU_LFC_TEMP              MU_MIXR          MU_CAPE_500 
#>              1.8e+01              1.5e+01              1.1e+03 
#>      MU_CAPE_500_M10   MU_CAPE_500_M10_PT           MU_CIN_500 
#>              3.9e+02              1.4e+02              0.0e+00 
#>            MU_LI_500        MU_LI_500_M10              SB_CAPE 
#>             -5.0e+00             -6.2e+00              2.3e+03 
#>          SB_CAPE_M10       SB_CAPE_M10_PT         SB_02km_CAPE 
#>              1.0e+03              3.1e+02              2.5e+02 
#>         SB_03km_CAPE          SB_HGL_CAPE               SB_CIN 
#>              5.8e+02              1.6e+03              0.0e+00 
#>           SB_LCL_HGT           SB_LFC_HGT            SB_EL_HGT 
#>              7.3e+02              7.3e+02              8.3e+03 
#>                SB_LI            SB_LI_M10              SB_WMAX 
#>             -1.0e+01             -1.1e+01              6.7e+01 
#>           SB_EL_TEMP          SB_LCL_TEMP          SB_LFC_TEMP 
#>             -2.9e+01              1.8e+01              1.8e+01 
#>              SB_MIXR              ML_CAPE          ML_CAPE_M10 
#>              1.5e+01              1.6e+03              6.7e+02 
#>       ML_CAPE_M10_PT         ML_02km_CAPE         ML_03km_CAPE 
#>              2.3e+02              1.6e+02              4.2e+02 
#>          ML_HGL_CAPE               ML_CIN           ML_LCL_HGT 
#>              1.3e+03              0.0e+00              9.8e+02 
#>           ML_LFC_HGT            ML_EL_HGT                ML_LI 
#>              9.8e+02              7.9e+03             -7.6e+00 
#>            ML_LI_M10              ML_WMAX           ML_EL_TEMP 
#>             -8.6e+00              5.7e+01             -2.6e+01 
#>          ML_LCL_TEMP          ML_LFC_TEMP              ML_MIXR 
#>              1.5e+01              1.5e+01              1.3e+01 
#>             LR_0500m              LR_01km              LR_02km 
#>             -1.0e+01             -1.0e+01             -1.0e+01 
#>              LR_03km              LR_04km              LR_06km 
#>             -9.0e+00             -7.9e+00             -6.7e+00 
#>              LR_16km              LR_26km              LR_24km 
#>             -6.0e+00             -5.0e+00             -5.8e+00 
#>              LR_36km          LR_26km_MAX         LR_500700hPa 
#>             -4.3e+00             -5.7e+00             -4.3e+00 
#>         LR_500800hPa         LR_600800hPa             FRZG_HGT 
#>             -5.2e+00             -5.8e+00              2.5e+03 
#>     FRZG_wetbulb_HGT  HGT_max_thetae_03km  HGT_min_thetae_04km 
#>              2.3e+03              0.0e+00              3.8e+03 
#>         Delta_thetae Delta_thetae_min04km          Thetae_01km 
#>              2.8e+01              2.9e+01              3.3e+02 
#>          Thetae_02km                DCAPE   Cold_Pool_Strength 
#>              3.2e+02              6.0e+02              1.3e+01 
#>           Wind_Index           PRCP_WATER   Moisture_Flux_02km 
#>              3.4e+01              2.7e+01              3.0e+01 
#>              RH_01km              RH_02km              RH_14km 
#>              7.3e-01              7.2e-01              6.5e-01 
#>              RH_25km              RH_36km               RH_HGL 
#>              5.6e-01              4.4e-01              4.6e-01 
#>             BS_0500m              BS_01km              BS_02km 
#>              1.9e+00              3.8e+00              8.8e+00 
#>              BS_03km              BS_06km              BS_08km 
#>              1.3e+01              1.8e+01              1.7e+01 
#>              BS_36km              BS_26km              BS_16km 
#>              9.4e+00              1.3e+01              1.7e+01 
#>              BS_18km            BS_EFF_MU            BS_EFF_SB 
#>              2.0e+01              1.4e+01              1.4e+01 
#>            BS_EFF_ML        BS_SFC_to_M10        BS_1km_to_M10 
#>              1.4e+01              1.6e+01              1.4e+01 
#>        BS_2km_to_M10     BS_MU_LFC_to_M10     BS_SB_LFC_to_M10 
#>              9.9e+00              1.4e+01              1.4e+01 
#>     BS_ML_LFC_to_M10        BS_MW02_to_SM        BS_MW02_to_RM 
#>              1.4e+01              7.3e+00              1.0e+01 
#>        BS_MW02_to_LM         BS_HGL_to_SM         BS_HGL_to_RM 
#>              1.1e+01              4.9e+00              7.8e+00 
#>         BS_HGL_to_LM             MW_0500m              MW_01km 
#>              1.0e+01              2.3e+00              2.4e+00 
#>              MW_02km              MW_03km              MW_06km 
#>              3.3e+00              4.8e+00              7.8e+00 
#>              MW_13km          SRH_100m_RM          SRH_250m_RM 
#>              6.8e+00              4.3e+00              1.0e+01 
#>          SRH_500m_RM           SRH_1km_RM           SRH_3km_RM 
#>              2.0e+01              4.0e+01              1.5e+02 
#>          SRH_36km_RM          SRH_100m_LM          SRH_250m_LM 
#>              2.4e+02              1.5e+00              3.6e+00 
#>          SRH_500m_LM           SRH_1km_LM           SRH_3km_LM 
#>              7.0e+00              1.4e+01             -1.3e+01 
#>          SRH_36km_LM           SV_500m_RM           SV_01km_RM 
#>             -2.4e+01              3.9e-03              3.9e-03 
#>           SV_03km_RM           SV_500m_LM           SV_01km_LM 
#>              4.8e-03              1.0e-03              1.1e-03 
#>           SV_03km_LM        MW_SR_500m_RM        MW_SR_01km_RM 
#>             -1.4e-03              1.0e+01              1.0e+01 
#>        MW_SR_03km_RM        MW_SR_500m_LM        MW_SR_01km_LM 
#>              9.5e+00              1.4e+01              1.3e+01 
#>        MW_SR_03km_LM     MW_SR_VM_500m_RM     MW_SR_VM_01km_RM 
#>              8.5e+00              1.0e+01              1.0e+01 
#>     MW_SR_VM_03km_RM     MW_SR_VM_500m_LM     MW_SR_VM_01km_LM 
#>              1.1e+01              1.4e+01              1.3e+01 
#>     MW_SR_VM_03km_LM       SV_FRA_500m_RM       SV_FRA_01km_RM 
#>              8.8e+00              1.0e+00              9.9e-01 
#>       SV_FRA_03km_RM       SV_FRA_500m_LM       SV_FRA_01km_LM 
#>              9.6e-01              2.6e-01              2.8e-01 
#>       SV_FRA_03km_LM         Bunkers_RM_A         Bunkers_RM_M 
#>             -2.9e-01              2.1e+02              7.8e+00 
#>         Bunkers_LM_A         Bunkers_LM_M         Bunkers_MW_A 
#>              1.2e+02              1.3e+01              1.5e+02 
#>         Bunkers_MW_M   Corfidi_downwind_A   Corfidi_downwind_M 
#>              7.8e+00              2.2e+02              1.5e+01 
#>     Corfidi_upwind_A     Corfidi_upwind_M              K_Index 
#>              2.3e+02              9.2e+00              2.4e+01 
#>      Showalter_Index    TotalTotals_Index          SWEAT_Index 
#>              3.8e+00              4.4e+01              1.1e+02 
#>              STP_fix              STP_new           STP_fix_LM 
#>              3.6e-01              2.0e-01              1.3e-01 
#>           STP_new_LM              SCP_fix              SCP_new 
#>              7.1e-02              6.2e+00              4.9e+00 
#>           SCP_fix_LM           SCP_new_LM                 SHIP 
#>             -5.4e-01             -4.2e-01              6.3e-01 
#>                  HSI                  DCP         MU_WMAXSHEAR 
#>              1.7e+00              1.2e+00              1.2e+03 
#>         SB_WMAXSHEAR         ML_WMAXSHEAR     MU_EFF_WMAXSHEAR 
#>              1.2e+03              1.0e+03              9.6e+02 
#>     SB_EFF_WMAXSHEAR     ML_EFF_WMAXSHEAR             EHI_500m 
#>              9.6e+02              8.0e+02              2.8e-01 
#>             EHI_01km             EHI_03km          EHI_500m_LM 
#>              5.6e-01              2.2e+00              9.9e-02 
#>          EHI_01km_LM          EHI_03km_LM              SHERBS3 
#>              2.0e-01             -1.9e-01              6.5e-01 
#>               SHERBE           SHERBS3_v2            SHERBE_v2 
#>              7.0e-01              8.6e-01              9.4e-01 
#>                  DEI              DEI_eff                  TIP 
#>              1.5e+00              1.2e+00              2.4e+00